[AE][ALSA] fixed period size calculations & better stream recovery
This fixes the earlier hack to calculate period and buffer sizes using hard coded values. We now use the same method as speaker-test.c which uses the maximum buffer size / 4 periords.

We also handle stream recovery better now producing less underrun errors and better A/V sync when an underrun occurs.

Note that there is still a HD underrun issue, but it appears to be originating in SoftAE as the stream still has frames when the underrun occurs.