[AE/PAP] Fixed numerous problems with PAP logic and sleep time calcs
The sleep time calculations could probably be written a little better but its almost 5am here :). It may fare better on IOS now, someone should remove that ifndef around the sleep code and test.

The addition of IsBuffering to the stream was required to ensure we dont calculate a sleep time on a stream that is buffering and not yet playing. It has the added benefit of allowing us to pre-buffer the streams before PAP starts their playback giving a smoother transition between gapless tracks and xfade